+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Add function `Direction::iter` which returns an iterator over all the `Direction` enum values
 - Add function `RoomXY::neighbors` which returns an iterator over all the valid neighbors of
   a given `RoomXY` position
+- Implement `JsCollectionIntoValue` and `JsCollectionFromValue` for `IntershardResourceType`
 
 0.21.0 (2024-05-14)
 ===================